# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nexus Academy API",
        "description": "Comprehensive OpenAPI 3.0 documentation for the Nexus Academy backend. This API handles user authentication, course management, enrollments, lessons, reviews, certificates, and instructor earnings. Authentication is handled via JWT tokens stored in cookies.",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@nexusacademy.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:4000/api",
            "description": "Local development server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "token",
                "description": "JWT token stored in a secure, httpOnly cookie named 'token'."
            }
        },
...
    },
    "security": [{ "cookieAuth": [] }],
    "paths": {
...
        "/certificates/download/{course_id}": {
            "get": {
                "tags": ["Certificates"],
                "summary": "Download certificate as PDF",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer" }
                    }
                ],
                "responses": { "200": { "description": "PDF" } }
            }
        },
...
    }
}